refactoring the code removing some unused images. Increasing the electron version to 30.1.0
