+++ app/appConfiguration/index.js
@@ -6,10 +6,7 @@ let _AppConfiguration_legacyConfigStore = new WeakMap();
 let _AppConfiguration_settingsStore = new WeakMap();
 
 class AppConfiguration {
-	/**
-	 * @param {string} configPath
-	 * @param appVersion
-	 */
+
 	constructor(configPath, appVersion) {
 		_AppConfiguration_configPath.set(this, configPath);
 		_AppConfiguration_startupConfig.set(this, require('../config')(configPath, appVersion));
@@ -23,9 +20,6 @@ class AppConfiguration {
 		}));
 	}
 
-	/**
-	 * @returns {string}
-	 */
 	get configPath() {
 		return _AppConfiguration_configPath.get(this);
 	}
@@ -34,16 +28,10 @@ class AppConfiguration {
 		return _AppConfiguration_startupConfig.get(this);
 	}
 
-	/**
-	 * @returns {Store}
-	 */
 	get legacyConfigStore() {
 		return _AppConfiguration_legacyConfigStore.get(this);
 	}
 
-	/**
-	 * @returns {Store}
-	 */
 	get settingsStore() {
 		return _AppConfiguration_settingsStore.get(this);
 	}
