+++ app/browser/index.js
@@ -57,10 +57,6 @@
 	window.Notification = CustomNotification;
 }());
 
-/**
- * @param {object} config 
- * @param {Electron.IpcRenderer} ipcRenderer 
- */
 function initializeModules(config, ipcRenderer) {
 	require('./tools/zoom').init(config);
 	require('./tools/shortcuts').init(config);
@@ -74,4 +70,3 @@ function initializeModules(config, ipcRenderer) {
 	require('./tools/customBackgrounds')(config, ipcRenderer);
 	require('./tools/theme').init(config, ipcRenderer);
 }
-
