+++ app/browser/notifications/activityManager.js
@@ -2,10 +2,6 @@ const activityHub = require('../tools/activityHub');
 const wakeLock = require('../tools/wakeLock');
 
 class ActivityManager {
-	/**
-	 * @param {Electron.IpcRenderer} ipcRenderer 
-	 * @param {./config} config
-	 */
 	constructor(ipcRenderer, config) {
 		this.ipcRenderer = ipcRenderer;
 		this.config = config;
