+++ app/browser/tools/customBackgrounds.js
@@ -5,15 +5,8 @@ let bgMSService = null;
 let bgMSMethod = null;
 // eslint-disable-next-line no-unused-vars
 let config = null;
-/**
- * @type {Electron.IpcRenderer}
- */
 let ipRenderer = null;
 
-/**
- * @param {object} conf 
- * @param {Electron.IpcRenderer} ipcr 
- */
 function init(conf, ipcr) {
 	config = conf;
 	ipRenderer = ipcr;
@@ -22,9 +15,6 @@ function init(conf, ipcr) {
 	});
 }
 
-/**
- * @param {{controller:object,injector:object}} inst 
- */
 async function overrideMSMethod(inst) {
 	bgMSService = inst.injector.get('customVideoBackgroundsService');
 	bgMSMethod = bgMSService.getProvidedImagesFromCdn;
@@ -32,9 +22,6 @@ async function overrideMSMethod(inst) {
 }
 
 async function customBGProvider(...args) {
-	/**
-	 * @type {Array<any>}
-	 */
 	const ms_response = config.customBGServiceIgnoreMSDefaults ? [] : await bgMSMethod.apply(bgMSService, [...args]);
 	const customList = await ipRenderer.invoke('getCustomBGList');
 	ms_response.push(...customList);
@@ -42,4 +29,3 @@ async function customBGProvider(...args) {
 }
 
 module.exports = init;
-
