+++ app/browser/tools/instance.js
@@ -1,7 +1,4 @@
 class Instance {
-	/**
-	 * @returns {Promise<{controller:object,injector:object}>}
-	 */
 	async whenReady(tries = 0) {
 		if (tries >= 5) {
 			throw new Error('Failed to get app objects after 5 tries');
@@ -33,4 +30,3 @@ async function sleep(ms) {
 }
 
 module.exports = new Instance();
-
