+++ app/browser/tools/mutationTitle.js
@@ -24,7 +24,6 @@ class MutationObserverTitle {
 		observer.observe(window.document.querySelector('title'),{ childList: true });
 		console.log('MutationObserverTitle logic applied');
 	}
-    
 }
 
 exports = module.exports = new MutationObserverTitle();
