+++ app/browser/tools/zoom.js
