+++ app/mainAppWindow/README.md
@@ -1,3 +1,5 @@
 # Main app window
 
 This folder contains the code to handle the main app browserWindow.
+
+The `windowManager.js` file manages the creation of windows and add the eventHandlers.
