+++ app/menus/index.js
@@ -12,9 +12,6 @@ const connectionManager = require('../connectionManager');
 let _Menus_onSpellCheckerLanguageChanged = new WeakMap();
 class Menus {
 	constructor(window, configGroup, iconPath) {
-		/**
-		 * @type {Electron.BrowserWindow}
-		 */
 		this.window = window;
 		this.iconPath = iconPath;
 		this.configGroup = configGroup;
@@ -25,16 +22,10 @@ class Menus {
 		this.initialize();
 	}
 
-	/**
-	 * @type {(languages:Array<string>)=>void}
-	 */
 	get onSpellCheckerLanguageChanged() {
 		return _Menus_onSpellCheckerLanguageChanged.get(this);
 	}
 
-	/**
-	 * @type {(languages:Array<string>)=>void}
-	 */
 	set onSpellCheckerLanguageChanged(value) {
 		if (typeof value === 'function') {
 			_Menus_onSpellCheckerLanguageChanged.set(this, value);
@@ -279,10 +270,6 @@ function assignAddToDictionaryHandler(params, menu, menus) {
 	addTextEditMenuItems(params, menu, menus);
 }
 
-/**
- * @param {Electron.Menu} menu 
- * @param {Menus} menus 
- */
 function addTextEditMenuItems(params, menu, menus) {
 	if (params.isEditable) {
 		buildEditContextMenu(menu, menus);
@@ -318,10 +305,6 @@ function buildEditContextMenu(menu, menus) {
 	addSpellCheckMenuItems(menu, menus);
 }
 
-/**
- * @param {Electron.Menu} menu 
- * @param {Menus} menus 
- */
 function addSpellCheckMenuItems(menu, menus) {
 	menu.append(
 		new MenuItem({
@@ -337,9 +320,6 @@ function addSpellCheckMenuItems(menu, menus) {
 	);
 }
 
-/**
- * @param {Menus} menus 
- */
 function createSpellCheckLanguagesMenu(menus) {
 	const activeLanguages = menus.window.webContents.session.getSpellCheckerLanguages();
 	const splChkMenu = new Menu();
@@ -361,10 +341,6 @@ function createSpellCheckLanguagesMenu(menus) {
 	return splChkMenu;
 }
 
-/**
- * @param {Electron.Menu} menu 
- * @param {Menus} menus 
- */
 function createSpellCheckLanguagesNoneMenuEntry(menu, menus) {
 	menu.append(
 		new MenuItem({
@@ -379,12 +355,6 @@ function createSpellCheckLanguagesNoneMenuEntry(menu, menus) {
 	);
 }
 
-/**
- * @param {{language:string,code:string}} language 
- * @param {Array<string>} activeLanguages 
- * @param {Menus} menus 
- * @returns 
- */
 function createLanguageMenuItem(language, activeLanguages, menus) {
 	return new MenuItem({
 		label: language.language,
@@ -395,10 +365,6 @@ function createLanguageMenuItem(language, activeLanguages, menus) {
 	});
 }
 
-/**
- * @param {MenuItem} item 
- * @param {Menus} menus 
- */
 function chooseLanguage(item, menus) {
 	const activeLanguages = menus.window.webContents.session.getSpellCheckerLanguages();
 	if (item) {
@@ -416,10 +382,6 @@ function chooseLanguage(item, menus) {
 	}
 }
 
-/**
- * @param {Array<string>} list 
- * @param {string} item 
- */
 function removeFromList(list, item) {
 	const itemIndex = list.findIndex(l => l == item);
 	if (itemIndex >= 0) {
@@ -429,10 +391,6 @@ function removeFromList(list, item) {
 	return list;
 }
 
-/**
- * @param {Array<string>} list 
- * @param {string} item 
- */
 function addToList(list, item) {
 	const itemIndex = list.findIndex(l => l == item);
 	if (itemIndex < 0) {
